Fix error when Imageset does not contain a mask image. Before, the pixel count was incorrect for process_image_ordered and tried to load a None when using the dict based extraction.

+++ trios/wop_matrix_ops.pxd
@@ -23,4 +23,4 @@ ctypedef fused raw_data:
     cython.uint
     cython.float
 
-cpdef long process_one_image(unsigned char[:,:] win, unsigned char[:,:] inp, unsigned char[:,:] out, unsigned char[:,:] msk, raw_data [:,:] X, unsigned char[:] y, FeatureExtractor extractor, raw_data [:] temp, int k)
+cpdef long process_one_image(unsigned char[:,:] win, unsigned char[:,:] inp, unsigned char[:,:] out, unsigned char[:,:] msk, raw_data [:,:] X, unsigned char[:] y, FeatureExtractor extractor, raw_data [:] temp, long k)
